backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
dj-database-url==2.3.0
psycopg2-binary==2.9.7
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-oauth-toolkit==1.7.1
python-keycloak==3.7.0
razorpay==1.4.1
Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.0
gunicorn==21.2.0
whitenoise==6.6.0
django-extensions==3.2.3
django-debug-toolbar==4.2.0
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2
flake8==6.1.0
black==23.11.0
isort==5.12.0
pre-commit==3.5.0
django-filter==23.5
drf-spectacular==0.26.5
django-rest-auth==0.9.5
django-allauth==0.57.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
django-environ==0.11.2
cryptography==41.0.7
PyJWT==2.8.0
requests==2.31.0
python-multipart==0.0.6
